+++ src/js/window.js
@@ -11,7 +11,7 @@ const dialog = remote.dialog;
 
 import TabGroup from 'electron-tabs';
 
-import Bookmark from './js/bookmark';
+import './js/bookmark_event';
 
 let tabGroup = null,
     addTabbutton = null,
@@ -30,11 +30,7 @@ let tabGroup = null,
     captureButton = null,
     captureResult = null,
     shiftKey = false,
-    cmdOrCtrlKey = false,
-    bookmarkSelect = null,
-    bookmarkMoveButton = null,
-    bookmarkDeleteButton = null,
-    bookmarkAddButton = null;
+    cmdOrCtrlKey = false;
 
 window.addEventListener('load', () => {
   tabGroup = new TabGroup();
@@ -57,11 +53,6 @@ window.addEventListener('load', () => {
   captureText = document.getElementById('capture-text');
   captureButton = document.getElementById('capture-button');
   captureResult = document.getElementById('capture-result');
-
-  bookmarkSelect = document.getElementById('bookmark-select');
-  bookmarkMoveButton = document.getElementById('bookmark-move-button');
-  bookmarkDeleteButton = document.getElementById('bookmark-delete-button');
-  bookmarkAddButton = document.getElementById('bookmark-add-button');
 
   addTabbutton.addEventListener('click', () => {
     createTab();
@@ -114,24 +105,6 @@ window.addEventListener('load', () => {
       captureFromUrls(captureText.value.split('\n'));
     }
   });
-
-  bookmarkMoveButton.addEventListener('click', () => {
-    const selected = bookmarkSelect.options[bookmarkSelect.selectedIndex];
-    if (selected) {
-      tabGroup.getActiveTab().webview.setAttribute('src', selected.value);
-    }
-  });
-  bookmarkDeleteButton.addEventListener('click', () => {
-    const selected = bookmarkSelect.options[bookmarkSelect.selectedIndex];
-    if (selected) {
-      deleteBookmark(selected.value);
-    }
-  });
-  bookmarkAddButton.addEventListener('click', () => {
-    addBookmark(tabGroup.getActiveTab().webview.src, tabGroup.getActiveTab().webview.getTitle());
-  })
-
-  prepareBookmarks();
 });
 
 window.addEventListener('keydown', (e) => {
@@ -337,43 +310,5 @@ async function request(url) {
     captureFromUrls(urls);
   } catch(error) {
     showDialog(error);
-  }
-}
-
-async function prepareBookmarks() {
-  for (let i = bookmarkSelect.options.length - 1; i >= 0; i--) {
-    bookmarkSelect.remove(i);
-  }
-
-  try {
-    // await Bookmark.clearAll();
-    const bookmarks = await Bookmark.getBookmarks();
-    const keys = Object.keys(bookmarks);
-    for (const key of keys) {
-      const option = document.createElement('option');
-      option.value = key;
-      option.text = `${bookmarks[key]}: ${key}`;
-      bookmarkSelect.add(option);
-    }
-  } catch(error) {
-    showDialog(error.toString());
-  }
-}
-
-async function deleteBookmark(url) {
-  try {
-    await Bookmark.deleteBookmark(url);
-    bookmarkSelect.remove(bookmarkSelect.selectedIndex);
-  } catch(error) {
-    showDialog(error.toString());
-  }
-}
-
-async function addBookmark(url, title) {
-  try {
-    await Bookmark.addBookmark(url, title);
-    prepareBookmarks();
-  } catch(error) {
-    showDialog(error.toString());
   }
 }
